+++ bignum/src/commonMain/kotlin/com/ionspin/kotlin/bignum/modular/ModularBigInteger.kt
@@ -100,6 +100,14 @@ class ModularBigInteger @ExperimentalUnsignedTypes constructor(
                     return ModularBigInteger(BigInteger.fromByte(byte).prep(), modulo, this)
                 }
 
+                override fun tryFromFloat(float: Float, exactRequired: Boolean): ModularBigInteger {
+                    return ModularBigInteger(BigInteger.tryFromFloat(float, exactRequired).prep(), modulo, this)
+                }
+
+                override fun tryFromDouble(double: Double, exactRequired: Boolean): ModularBigInteger {
+                    return ModularBigInteger(BigInteger.tryFromDouble(double, exactRequired).prep(), modulo, this)
+                }
+
                 private fun BigInteger.prep() : BigInteger {
                     val result = this % modulo
                     return when (result.sign) {