+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## Descriptive changelog
 (All dates are DD.MM.YYYY)
 
-##### 0.1.0-SNAPSHOT - 1.6.2019
+##### 0.1.0-SNAPSHOT - 2.6.2019
 - Added toFloat and toDouble to BigInteger and ModularBigInteger classes
+- Added BigInteger creation from Float and Double by using `tryFromFloat` and `tryFromDouble`, with optional exact 
+parameter to preserve precision.
+- Added BigInteger comparison with Float and Double
 
 ##### 0.0.9 - 11.5.2019 Adding modular integer support, changing api
 - Added modular integers - ModularBigInteger